Separate actions from store.
After more discussion with Evan, we decided that actions don't really belong in the store modules.

The reason is that there isn't a one-to-one correspondence between the action services and the reducers; indeed actions allow us to decouple our view from our store.

The store modules should focus on reducers and helpers for the reducers (types, serializations).